% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabarito_carrega_dados.R
\name{carrega_cnep}
\alias{carrega_cnep}
\title{Carrega dados do CNEP}
\usage{
carrega_cnep(ano_inicial = 2015, ano_final = 2016)
}
\arguments{
\item{ano_inicial}{Ano em que se inicia o intervalo de busca por sanções}

\item{ano_final}{Ano em que se finda o intervalo de busca por sanções}
}
\value{
Dataframe com dados do CNEP
}
\description{
Carrega do Portal da Transparência os dados de empresas presentes no CNEP cujas sanções foram aplicadas em um intervalo especificado.
O dataframe retornado conta com as seguintes variáveis:
    \describe{
        \item{cd_Sancionado}{CPF ou CNPJ do sancionado}
        \item{no_Sancionado}{Nome do sancionado}
        \item{dt_Inicio}{Data de início da sanção}
        \item{dt_Final}{Data do final da sanção}
    }
}
