% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_fornecedores_pagamentos.R
\name{carrega_contratos}
\alias{carrega_contratos}
\title{Carrega contratos}
\usage{
carrega_contratos(ano_inicial = 2014, ano_final = 2014,
  limite_inferior = 140000)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo}

\item{ano_final}{Ano final do intervalo de tempo}

\item{limite_inferior}{Limite inferior para o valor do contrato}
}
\value{
Data frame com informações do contrato, como o contratado e o valor do contrato.
}
\description{
Carrega contratos realizados em um intervalo de tempo
O dataframe retornado conta com as seguintes variáveis:
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{dt_Ano}{Ano em que o empenho foi realizado}
        \item{nu_Contrato}{Número do contrato}
        \item{dt_Assinatura}{Data da assinatura dos contratos}
        \item{pr_Vigencia}{Prazo de vigência do contrato}
        \item{nu_CPFCNPJ}{Código do credor (CPF, CNPJ)}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitação}{Tipo da licitação}
        \item{vl_TotalContrato}{Valor total do contrato}
        \item{dt_MesAno}{Mês e ano em que o contrato foi firmado}
    }
}
