% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_empenhos.R
\name{carrega_empenhos}
\alias{carrega_empenhos}
\title{Carrega empenhos}
\usage{
carrega_empenhos(ano_inicial = 2014, ano_final = 2014, lista_cnpjs)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo}

\item{ano_final}{Ano final do intervalo de tempo}

\item{lista_cnpjs}{Lista com os CNPJ's dos credores de interesse}
}
\value{
Data frame com informações do empenho, do valor de pagamentos associados a este empenho e o valor estornado associado aos pagamentos
}
\description{
Carrega a lista de empenhos destinado a um grupo de credores em um intervalo de tempo.
    O dataframe retornado conta com as seguintes variáveis:
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{dt_Ano}{Ano em que o empenho foi realizado}
        \item{cd_UnidOrcamentaria}{Unidade Orcamentária de origem}
        \item{nu_Empenho}{Número do empenho}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{cd_Credor}{Código do credor (CPF, CNPJ)}
        \item{no_Credor}{Nome do credor}
        \item{vl_Pago}{Valor dos pagamentos associados ao empenho}
        \item{vl_Estornado}{Valor dos estornos associados ao empenho (quando houver)}
    }
}
