% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_empenhos.R
\name{carrega_empenhos_data_limite}
\alias{carrega_empenhos_data_limite}
\title{carrega_empenhos_data_limite}
\usage{
carrega_empenhos_data_limite(ano_inicial = 2014, ano_final = 2014,
  cnpjs_datas_contratos)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo}

\item{ano_final}{Ano final do intervalo de tempo}

\item{cnpjs_datas_contratos}{Dataframe de CNPJ's de interesse e datas de início de contrato para o cálculo das informações de fornecimento (nu_CPFCNPJ, dt_Inicio)}
}
\value{
Data frame com informações do empenho, do valor de pagamentos associados a este empenho e o valor estornado associado aos pagamentos
}
\description{
Carrega a lista de empenhos realizados em um intervalo de tempo.
    O dataframe retornado conta com as seguintes variáveis.
    \describe{
        \item{dt_Contrato}{Data em que o contrato foi firmado}
        \item{cd_Credor}{Código do credor (CPF, CNPJ)}
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{dt_Ano}{Ano em que o empenho foi realizado}
        \item{cd_UnidOrcamentaria}{Unidade Orcamentária de origem}
        \item{nu_Empenho}{Número do empenho}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{no_Credor}{Nome do credor}
        \item{vl_Pago}{Valor dos pagamentos associados ao empenho}
        \item{vl_Estornado}{Valor dos estornos associados ao empenho (quando houver)}
    }
}
