% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gabarito_carrega_dados.R
\name{carrega_gabarito_sancoes}
\alias{carrega_gabarito_sancoes}
\title{carrega_gabarito_sancoes}
\usage{
carrega_gabarito_sancoes(data_versao = format(Sys.time(), "\%Y\%m\%d"),
  ano_inicial = 2015, ano_final = 2016)
}
\arguments{
\item{data_versao}{A versão dos dados que deve ser obtida do Portal da Transparência no formato YYYYmmdd (As atualizações são diárias)}

\item{ano_inicial}{Ano em que se inicia o intervalo de busca por sanções}

\item{ano_final}{Ano em que se finda o intervalo de busca por sanções}
}
\value{
Dataframe com dados do CNEP e do CEIS
}
\description{
Carrega dados de empresas presentes no CNEP e no CEIS cujas sanções foram aplicadas em um intervalo especificado.
O dataframe retornado conta com as seguintes variáveis:
    \describe{
        \item{cd_Sancionado}{CPF ou CNPJ do sancionado}
        \item{no_Sancionado}{Nome do sancionado}
        \item{dt_Inicio}{Data de início da sanção}
        \item{dt_Final}{Data do final da sanção}
        \item{tp_Origem}{Origem da sanção (CNEP, CEIS)}
    }
}
