% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_info_contrato.R
\name{carrega_info_contrato}
\alias{carrega_info_contrato}
\title{Carrega informações de um contrato}
\usage{
carrega_info_contrato(db_user, dados_contrato, ano_inicial = 2011,
  ano_final = 2016, limite_inferior = 0)
}
\arguments{
\item{db_user}{Usuario utilizado para conexão no BD}

\item{dados_contrato}{Informações dos contratos com informando o valor total do contrato e o montante recebido por ano (cd_UGestora, nu_Contrato, nu_CPFCNPJ, dt_Inicio, vl_Contrato, vl_GanhoAnual)}

\item{ano_inicial}{Ano inicial do intervalo de tempo (limite inferior). Default é 2011.}

\item{ano_final}{Ano final do intervalo de tempo (limite superior). Default é 2016.}

\item{limite_inferior}{Menor valor de contrato a ser considerado ao calcular as tipologias}
}
\value{
Data frame com informações do contrato. Ex: Razão entre o valor do contrato e o montante recebido
}
\description{
Carrega a informações de contratos associadas a CNPJ's
   \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Contrato}{Número do contrato}
        \item{nu_CPFCNPJ}{Código do credor (CPF, CNPJ)}
        \item{dt_Início}{Data em que o empenho foi realizado}
        \item{pr_ContratoTotalGanho}{Proporção do valor do contrato sobre o valor total ganho pelo credor em todos contratos até a assinatura}
        \item{md_Contratos}{Valor médio obtido pelo credor em todos os contratos realizados}
        \item{qt_Contratos}{Quantidade de contratos firmados pelo credor}
    }
}
