% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_info_fornecimento.R
\name{carrega_info_fornecimento}
\alias{carrega_info_fornecimento}
\title{Carrega informações de fornecimento}
\usage{
carrega_info_fornecimento(ano_inicial, ano_final, cnpjs_datas_contratos)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo}

\item{ano_final}{Ano final do intervalo de tempo}

\item{cnpjs_datas_contratos}{Lista de CNPJ's e de datas de início de contrato para o cálculo das informações de fornecimento.}
}
\value{
Data frame com informações sobre as empresas com uma data de limite superior como número de municípios com fornecimento, total de dinheiro recebido.
}
\description{
Calcula informações sobre o fornecimento de uma lista de CNPj's.
   \describe{
        \item{cd_Credor}{Código do credor (CPF, CNPJ)}
        \item{no_Credor}{Nome do credor}
        \item{dt_Inicio}{Data em que o contrato foi assinado}
        \item{md_Municipios}{Número médio anual de municípios para os quais o credor forneceu serviços até a assinatura do contrato}
        \item{qt_Municipios}{Número total de municípios para os quais o credor forneceu serviços até a assinatura do contrato}
        \item{md_UGestoras}{Número médio anual de unidades gestoras para os quais o credor forneceu serviços até a assinatura do contrato}
        \item{qt_UGestoras}{Número total de unidades gestoras para os quais o credor forneceu serviços até a assinatura do contrato}
        \item{md_Ganho}{Valor médio anual ganho pelo credor até a assinatura do contrato}
        \item{vl_Ganho}{Valor ganho pelo credor até a assinatura do contrato}
    }
}
