% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_info_licitacao.R
\name{carrega_info_licitacao}
\alias{carrega_info_licitacao}
\title{Carrega informações de licitações}
\usage{
carrega_info_licitacao(ano_inicial = 2011, ano_final = 2016,
  cnpjs_datas_contratos)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo (limite inferior). Default é 2011.}

\item{ano_final}{Ano final do intervalo de tempo (limite superior). Default é 2016.}

\item{cnpjs_datas_contratos}{Lista de CNPJ's e de datas de início de contrato para o cálculo das informações de licitação}
}
\value{
Data frame com informações da licitação. Para cada empresa (CNPJ) informações sobre a quantidade de licitações e o montante de dinheiro envolvido.
}
\description{
Carrega a informações de licitações associadas a CNPJ's participantes da licitação.
    O dataframe retornado conta com as seguintes variáveis:
   \describe{
        \item{nu_CPFCNPJ}{Código do credor (CPF, CNPJ)}
        \item{dt_Inicio}{Data em que o contrato foi iniciado}
        \item{qt_LicitacoesParticipou}{Quantidade de licitações em que o credor participou até a data de um contrato}
        \item{qt_LicitacoesVenceu}{Quantidade de licitações em que o credor venceu até a data de um contrato}
        \item{vl_TotalLicitacoesVenceu}{O valor total de todas as licitações vencidas pelo credor até a data de um contrato}
        \item{pr_Vitorias}{A razão entre a quantidade de vitórias e a quantidade de participações do credor até a data de um contrato}
        \item{md_LicitacoesParticipou}{Quantidade média anual de licitações em que o credor participou até a data de um contrato}
        \item{md_LicitacoesVenceu}{Quantidade média anual de licitações em que o credor venceu até a data de um contrato}
    }
}
