% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_info_proposta.R
\name{carrega_info_proposta}
\alias{carrega_info_proposta}
\title{Carrega informações de propostas}
\usage{
carrega_info_proposta(db_user, ano_inicial = 2011, ano_final = 2016,
  cnpjs_datas_contratos)
}
\arguments{
\item{db_user}{Usuario utilizado para conexão no BD}

\item{ano_inicial}{Ano inicial do intervalo de tempo (limite inferior). Default é 2011.}

\item{ano_final}{Ano final do intervalo de tempo (limite superior). Default é 2016.}

\item{cnpjs_datas_contratos}{Lista de CNPJ's e de datas de início de contrato para o cálculo das informações das propostas.}
}
\value{
Data frame com informações de quantas propostas foram feitas por uma determinada empresa até uma data limite.
}
\description{
Carrega a informações do número de propostas de CNPJS associadas a licitações dentro de um intervalo de tempo.
    Retorna um dataframe com as seguintes variáveis:
    \describe{
        \item{nu_CPFCNPJ}{Código do credor (CPF, CNPJ)}
        \item{dt_Inicio}{Data em que o contrato foi iniciado}
        \item{qt_Propostas}{Quantidade de propostas feitas pela empresa até a assinatura do contrato}
        \item{md_Propostas}{Quantidade média de propostas feitas por ano pela empresa até a assinatura do contrato}
    }
}
