% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_licitacoes.R
\name{carrega_licitacoes}
\alias{carrega_licitacoes}
\title{Carrega licitações}
\usage{
carrega_licitacoes(ano_inicial = 2011, ano_final = 2016)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo (limite inferior). Default é 2011.}

\item{ano_final}{Ano final do intervalo de tempo (limite superior). Default é 2016.}
}
\value{
Data frame com informações da licitação. Para cada empresa (CNPJ) informações sobre a quantidade de licitações e o montante de dinheiro envolvido por ano
}
\description{
Carrega a informações de licitações associadas a CNPJ's participantes da licitação.
    O dataframe retornado conta com as seguintes variáveis.
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{dt_Ano}{Ano em que a licitação foi realizada}
        \item{dt_Homologacao}{Data em que a licitação foi homologada}
        \item{vl_Licitacao}{Valor da licitação}
    }
}
