% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_fornecedores_pagamentos.R
\name{carrega_licitacoes_por_fornecedor}
\alias{carrega_licitacoes_por_fornecedor}
\title{carrega_licitacoes_por_fornecedor}
\usage{
carrega_licitacoes_por_fornecedor(ano_inicial = 2014, ano_final = 2014,
  lista_cnpjs)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo}

\item{ano_final}{Ano final do intervalo de tempo}

\item{lista_cnpjs}{Lista com os CNPJ's dos credores de interesse}
}
\value{
Data frame com informações da licitação, incluindo nome do credor vitorioso, total pago, total estornado e número de empenhos
}
\description{
Carrega a lista de licitações por fornecedor com informações de pagamentos em um intervalo de tempo.
O dataframe retornado conta com as seguintes variáveis:
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{dt_Ano}{Ano em que o empenho foi realizado}
        \item{cd_Credor}{Código do credor (CPF, CNPJ)}
        \item{no_Credor}{Nome do credor}
        \item{qt_Empenhos}{Quantidade de empenhos destinados ao credor gerados por uma licitação}
        \item{vl_Pago}{Valor dos pagamentos associados ao empenho}
        \item{vl_Estornado}{Valor dos estornos associados ao empenho (quando houver)}
        \item{vl_Ganho}{Valor do ganho total associado ao empenho (quando houver)}
    }
}
