% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_licitacoes.R
\name{carrega_propostas_licitacao}
\alias{carrega_propostas_licitacao}
\title{carrega_propostas_licitacao}
\usage{
carrega_propostas_licitacao(ano_inicial = 2011, ano_final = 2016)
}
\arguments{
\item{ano_inicial}{Ano inicial do intervalo de tempo (limite inferior). Default é 2011.}

\item{ano_final}{Ano final do intervalo de tempo (limite superior). Default é 2016.}
}
\value{
Data frame com informações das propostas para licitações em um intervalo de tempo
}
\description{
Carrega a informações das propostas associadas a licitações que ocorreram dentro de um intervalo de tempo
    O dataframe retornado conta com as seguintes variáveis.
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{dt_Homologacao}{Data em que a licitação foi homologada}
        \item{dt_Ano}{Ano em que a licitação foi homologada}
        \item{cd_Item}{Código do item ao qual se refere uma proposta}
        \item{cd_SubGrupoItem}{Código do subgrupo do item ao qual se refere uma proposta}
        \item{vl_Ofertado}{Valor ofertado pelo objeto da licitação}
    }
}
