% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_empenhos.R
\name{empenhos_credores_data}
\alias{empenhos_credores_data}
\title{Carrega informações dos credores}
\usage{
empenhos_credores_data(data_limite_superior, empenhos, pagamentos,
  estornos_pagamento, cnpjs_datas_contratos)
}
\arguments{
\item{data_limite_superior}{Data limite para a data do pagamento e do estorno}

\item{empenhos}{Data frame com todos os empenhos recuperados do SAGRES}

\item{pagamentos}{Data frame com todos os pagamentos recuperados do SAGRES}

\item{estornos_pagamento}{Data frame com todos os estornos de pagamentos recuperados do SAGRES}

\item{cnpjs_datas_contratos}{Lista de CNPJ's e de datas de início de contrato para o cálculo das informações de fornecimento.}
}
\value{
Data frame com informações do empenho, do valor de pagamentos associados a este empenho e o valor estornado associado aos pagamentos
}
\description{
Cruza informações sobre empenhos, pagamentos e estornos a partir de uma data de limite.
Retorna todos os empenhos até esta data num dataframe com as variáveis:
   \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{dt_Ano}{Ano em que o empenho foi realizado}
        \item{cd_UnidOrcamentaria}{Unidade Orcamentária de origem}
        \item{nu_Empenho}{Número do empenho}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{cd_Credor}{Código do credor (CPF, CNPJ)}
        \item{no_Credor}{Nome do credor}
        \item{vl_Pago}{Valor dos pagamentos associados ao empenho}
        \item{vl_Estornado}{Valor dos estornos associados ao empenho (quando houver)}
    }
}
