% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{get_empenhos_licitacao_filt}
\alias{get_empenhos_licitacao_filt}
\title{get_empenhos_licitacao_filt}
\usage{
get_empenhos_licitacao_filt(dbcon, nu_licitacao, cd_ugestora, tp_licitacao,
  cd_funcao, cd_subfuncao, cd_subelemento = 99)
}
\arguments{
\item{dbcon}{conexão com o banco de dados}

\item{nu_licitacao}{o número da licitação cujo percentual de aditivos deve ser calculado}

\item{cd_ugestora}{o código da unidade gestora que lançou a licitação em questão}

\item{tp_licitacao}{o tipo da licitação em questão}

\item{cd_funcao}{código de função a ser utilizado na filtragem}

\item{cd_subfuncao}{código de subfunção a ser utilizado na filtragem}

\item{cd_subelemento}{código de subelemento a ser utilizado na filtragem. Valor padrão: 99 (Sem subelemento)}
}
\description{
Retorna os empenhos da licitacao que corresponde aos códigos passados como parâmetro
}
