% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{get_licitacoes}
\alias{get_licitacoes}
\title{get_licitacoes}
\usage{
get_licitacoes(dbcon, cd_funcao, cd_subfuncao, cd_subelemento)
}
\arguments{
\item{dbcon}{conexão com o banco de dados}

\item{cd_funcao}{código de função (integer)}

\item{cd_subfuncao}{código de subfunção (integer)}

\item{cd_subelemento}{código de subelemento (character)}
}
\description{
recupera as licitações cujos empenhos apresentam todos código de função,
  código de subfunção e código de subelemento iguais aos fornecidos
}
