% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{get_liquidacoes_filtradas}
\alias{get_liquidacoes_filtradas}
\title{get_liquidacoes_filtradas}
\usage{
get_liquidacoes_filtradas(dbcon, cd_funcao, cd_subfuncao, cd_subelemento = 99)
}
\arguments{
\item{dbcon}{conexão com o banco de dados}

\item{cd_funcao}{código de função a ser utilizado na filtragem}

\item{cd_subfuncao}{código de subfunção a ser utilizado na filtragem}

\item{cd_subelemento}{código de subelemento a ser utilizado na filtragem. Valor padrão: 99 (Sem subelemento)}
}
\description{
Filtra as liquidações realizadas no estado da Paraíba a partir do 
código de função, código de subfunção e código de subelemento dos empenhos a que elas estão relacionadas.
}
