% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{get_percentual_aditivos}
\alias{get_percentual_aditivos}
\title{get_percentual_aditivos}
\usage{
get_percentual_aditivos(dbcon, nu_licitacao, cd_ugestora, tp_licitacao)
}
\arguments{
\item{dbcon}{conexão com o banco de dados}

\item{nu_licitacao}{o número da licitação cujo percentual de aditivos deve ser calculado}

\item{cd_ugestora}{o código da unidade gestora que lançou a licitação em questão}

\item{tp_licitacao}{o tipo da licitação em questão}
}
\description{
Calcula o percentual do valor de aditivos que incide sobre uma licitação.
Um percentual de exatamente 0 indica que provavelmente o aditivo solicitado para a licitação não foi monetário, mas sim de tempo.
Um retorno 'empty' indica que não há aditivos cadastrados para a licitação analisada.
ATENÇÃO: Não é necessário realizar o collect do retorno deste método.
}
