% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{get_valor_empenhos_licitacoes}
\alias{get_valor_empenhos_licitacoes}
\title{get_valor_empenhos_licitacoes}
\usage{
get_valor_empenhos_licitacoes(dbcon, cd_funcao, cd_subfuncao,
  cd_subelemento = 99)
}
\arguments{
\item{dbcon}{conexão com o banco de dados}

\item{cd_funcao}{código de função a ser utilizado na filtragem}

\item{cd_subfuncao}{código de subfunção a ser utilizado na filtragem}

\item{cd_subelemento}{código de subelemento a ser utilizado na filtragem. Valor padrão: 99 (Sem subelemento)}
}
\description{
Retorna o valor total dos empenhos que correspondem aos códigos passados como parâmetro em todas as licitações
}
