% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_valida_info_fornecimento_credor.R
\name{info_credor_data}
\alias{info_credor_data}
\title{Carrega informações de credores}
\usage{
info_credor_data(cd_Credor, dt_Inicio)
}
\arguments{
\item{cd_Credor}{Código do Credor (CPF ou CNPJ)}

\item{dt_Inicio}{Data limite para a data do pagamento e do estorno}
}
\value{
Lista com dois data frames: um contendo as informações do total pago por ano e outro contendo informações do total estornado por ano
}
\description{
Recupera informações de empenhos a partir de um CPF/CNPJ e uma data limite. As informações são o total pago e o total estornado
}
