% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_aditivos_merenda}
\alias{load_aditivos_merenda}
\title{Carrega os aditivos de merenda realizados no estado da Paraíba.
Os aditivos de merenda são os aditivos atrelados a licitações
classificadas como licitações de merenda. Para saber mais sobre
o método de filtragem de licitações, consulte \code{\link{load_licitacoes_merenda}}}
\usage{
load_aditivos_merenda()
}
\description{
Carrega os aditivos de merenda realizados no estado da Paraíba.
Os aditivos de merenda são os aditivos atrelados a licitações
classificadas como licitações de merenda. Para saber mais sobre
o método de filtragem de licitações, consulte \code{\link{load_licitacoes_merenda}}
}
