% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_contratos_merenda}
\alias{load_contratos_merenda}
\title{Carrega os contratos de merenda realizados no estado da Paraíba.
Os contratos de merenda são os contratos atrelados a licitações
classificadas como licitações de merenda. Para saber mais sobre
o método de filtragem de licitações, consulte \code{\link{load_licitacoes_merenda}}}
\usage{
load_contratos_merenda()
}
\description{
Carrega os contratos de merenda realizados no estado da Paraíba.
Os contratos de merenda são os contratos atrelados a licitações
classificadas como licitações de merenda. Para saber mais sobre
o método de filtragem de licitações, consulte \code{\link{load_licitacoes_merenda}}
}
