% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dados_escolares}
\alias{load_dados_escolares}
\title{Carrega varios arquivos com dados escolares de todas as cidades da Paraíba
e os agrupa em um único arquivo}
\usage{
load_dados_escolares(caminho_arquivo)
}
\arguments{
\item{caminho_arquivo}{O caminho onde estão localizados os arquivos csv a serem processados}
}
\value{
Um único arquivo csv contendo dados escolares de todos os municípios da Paraíba
}
\description{
Carrega varios arquivos com dados escolares de todas as cidades da Paraíba
e os agrupa em um único arquivo
}
