% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_empenhos_merenda}
\alias{load_empenhos_merenda}
\title{Carrega os empenhos de merenda realizadas no estado da Paraíba.
Um empenho é considerado um empenho de merenda quando possui função
de 'educação' e subelemento 'gênero alimentício' ou quando possui função
de 'educação' e subfunção de 'alimentação e nutrição'}
\usage{
load_empenhos_merenda()
}
\description{
Carrega os empenhos de merenda realizadas no estado da Paraíba.
Um empenho é considerado um empenho de merenda quando possui função
de 'educação' e subelemento 'gênero alimentício' ou quando possui função
de 'educação' e subfunção de 'alimentação e nutrição'
}
