% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_licitacoes_merenda}
\alias{load_licitacoes_merenda}
\title{Carrega as licitacoes de merenda realizadas no estado da Paraíba.
Uma licitação é considerada uma licitação de merenda se pelo menos
80% dos seus empenhos é filtrado como um empenho de merenda.
O método de filtragem de empenhos pode ser visto na documentação de
\code{\link{load_empenhos_merenda}}}
\usage{
load_licitacoes_merenda()
}
\description{
Carrega as licitacoes de merenda realizadas no estado da Paraíba.
Uma licitação é considerada uma licitação de merenda se pelo menos
80% dos seus empenhos é filtrado como um empenho de merenda.
O método de filtragem de empenhos pode ser visto na documentação de
\code{\link{load_empenhos_merenda}}
}
