% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_licitantes_merenda}
\alias{load_licitantes_merenda}
\title{Carrega informações sobre os licitantes de merenda do estado da Paraíba}
\usage{
load_licitantes_merenda()
}
\description{
Carrega informações sobre os licitantes de merenda do estado da Paraíba
}
