% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_liquidacoes_merenda}
\alias{load_liquidacoes_merenda}
\title{Carrega as liquidações de merenda realizadas no estado da Paraíba.
As liquidações de merenda são as liquidações efetuadas em empenhos
classificados como empenhos de merenda. Para saber mais sobre o 
método de filtragem de empenhos, consulte \code{\link{load_empenhos_merenda}}}
\usage{
load_liquidacoes_merenda()
}
\description{
Carrega as liquidações de merenda realizadas no estado da Paraíba.
As liquidações de merenda são as liquidações efetuadas em empenhos
classificados como empenhos de merenda. Para saber mais sobre o 
método de filtragem de empenhos, consulte \code{\link{load_empenhos_merenda}}
}
