% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsagrespb.R
\name{load_pagamentos_merenda}
\alias{load_pagamentos_merenda}
\title{Carrega os pagamentos de merenda realizados no estado da Paraíba.
Os pagamentos de merenda são os pagamentos efetuados em empenhos
classificados como empenhos de merenda. Para saber mais sobre o 
método de filtragem de empenhos, consulte \code{\link{load_empenhos_merenda}}}
\usage{
load_pagamentos_merenda()
}
\description{
Carrega os pagamentos de merenda realizados no estado da Paraíba.
Os pagamentos de merenda são os pagamentos efetuados em empenhos
classificados como empenhos de merenda. Para saber mais sobre o 
método de filtragem de empenhos, consulte \code{\link{load_empenhos_merenda}}
}
