% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrega_licitacoes.R
\name{carrega_licitacoes_vencedores}
\alias{carrega_licitacoes_vencedores}
\title{carrega_licitacoes_vencedores}
\usage{
carrega_licitacoes_vencedores(ano_inicial = 2011, ano_final = 2016)
}
\arguments{
\item{ano_inicial}{Ano inicial do período de tempo}

\item{ano_final}{Ano final do período de tempo}
}
\value{
Data frame com informações da licitação e de seus vencedores
}
\description{
Carrega lista de licitações realizadas em um determinado período de tempo
    O dataframe retornado conta com as seguintes variáveis.
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{cd_Credor}{Código do credor (CPF/CNPJ)}
        \item{min_dt_Empenho}{Data do primeiro empenho}
    }
}
