% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagres_carrega_licitacoes.R
\name{carrega_participantes}
\alias{carrega_participantes}
\title{Carrega participantes de licitações}
\usage{
carrega_participantes(db_user, lista_cnpjs)
}
\arguments{
\item{db_user}{Usuario utilizado para conexão no BD}

\item{lista_cnpjs}{Lista de CNPJ's que se quer pesquisar}
}
\value{
Data frame com informações de participações em licitação para cada CNPJ da lista de CNPJ's.
}
\description{
Carrega informações de participação em licitações a partir de uma lista de CNPJ's
    O dataframe retornado conta com as seguintes variáveis.
    \describe{
        \item{cd_UGestora}{Unidade Gestora de origem}
        \item{nu_Licitacao}{Número da licitação}
        \item{tp_Licitacao}{Tipo da licitação}
        \item{dt_Ano}{Ano em que a licitação foi realizada}
        \item{dt_MesAno}{Data em que a licitação foi homologada (Mês e ano)}
        \item{nu_CPFCNPJ}{Código do credor (CPF/CNPJ)}
    }
}
