% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dados_escolares}
\alias{load_dados_escolares}
\title{Carrega os dados escolares de todas as cidades da Paraíba}
\usage{
load_dados_escolares(caminho_arquivo)
}
\arguments{
\item{caminho_arquivo}{O caminho onde estão localizados os arquivos csv a serem processados}
}
\value{
Um único arquivo csv contendo dados escolares de todos os municípios da Paraíba
}
\description{
Carrega os dados escolares de todas as cidades da Paraíba
}
